+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v0.1.229
+
 - Rest\Query\Filter: fix mapConditionNodes for ConstantNode nodes
 - Rest\Query\Filter: add operator type 'HAS' to check if an array contains a certain value